kernel/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.4
paho-mqtt==1.6.1
pymongo==4.6.0
requests==2.31.0
urllib3==2.1.0
